Added Responsive features
