+++ lib/features/cartc/widgets/cart_listview.dart
@@ -8,72 +8,97 @@ class CartListView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      height: MediaQuery.sizeOf(context).height * 0.22,
-      child: ListView(
-        scrollDirection: Axis.horizontal,
-        children: [
-          Container(
-            width: MediaQuery.sizeOf(context).width * 0.36,
-            decoration: BoxDecoration(
-              color: Colors.blue.shade50,
-              border: Border.all(color: Color(0xFF0F26A6), width:2),
-              borderRadius: BorderRadius.circular(16),
+    return Container(
+     // color: Colors.red,
+      child: SizedBox(
+        height: MediaQuery.of(context).orientation == Orientation.portrait
+        ?  MediaQuery.sizeOf(context).height * 0.15
+            : MediaQuery.sizeOf(context).height * 0.3,
+        child: ListView(
+          scrollDirection: Axis.horizontal,
+          children: [
+            Container(
+              width: MediaQuery.sizeOf(context).width * 0.25,
+              //height: MediaQuery.sizeOf(context).height * 0.005,
+              decoration: BoxDecoration(
+                color: Colors.blue.shade50,
+                border: Border.all(color: Color(0xFF0F26A6), width:2),
+                borderRadius: BorderRadius.circular(16),
+              ),
+
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Image.asset("assets/images/cart-images/btn-shirt.png", width:50,),
+                  SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
+                  Text("Wash and Fold", style: TextStyle(color: Color(0xFF0F26A6), fontSize: 10),),
+                ],
+              ),
             ),
 
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Image.asset("assets/images/cart-images/btn-shirt.png", width:80,),
-                SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
-                Text("Wash and Fold", style: TextStyle(color: Color(0xFF0F26A6)),),
-              ],
+            SizedBox (width: MediaQuery.sizeOf(context).width * 0.02,),
+            Container(
+              width: MediaQuery.sizeOf(context).width * 0.25,
+              decoration: BoxDecoration(
+                color: Colors.blue.shade50,
+                //border: Border.all(color: Color(0xFF0F26A6), width:2),
+                borderRadius: BorderRadius.circular(16),
+              ),
+
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Image.asset("assets/images/cart-images/washing-machine.png", width: 40,),
+                  SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
+                  Text("Dry Cleaning", style: TextStyle(color: Color(0xFF0F26A6), fontSize: 10),),
+
+
+                ],
+              ),
             ),
-          ),
+            SizedBox (width: MediaQuery.sizeOf(context).width * 0.02,),
+            Container(
+              width: MediaQuery.sizeOf(context).width * 0.25,
+              decoration: BoxDecoration(
+                color: Colors.blue.shade50,
+                //border: Border.all(color: Color(0xFF0F26A6), width:2),
+                borderRadius: BorderRadius.circular(16),
+              ),
 
-          SizedBox (width: MediaQuery.sizeOf(context).width * 0.05,),
-          Container(
-            width: MediaQuery.sizeOf(context).width * 0.36,
-            decoration: BoxDecoration(
-              color: Colors.blue.shade50,
-              //border: Border.all(color: Color(0xFF0F26A6), width:2),
-              borderRadius: BorderRadius.circular(16),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Image.asset("assets/images/cart-images/handwashing.png", width: 40,),
+                  SizedBox(height: MediaQuery.sizeOf(context).height * 0.03,),
+                  Text("Hand Washing", style: TextStyle(color: Color(0xFF0F26A6), fontSize: 10),),
+
+
+                ],
+              ),
+            ),
+            SizedBox (width: MediaQuery.sizeOf(context).width * 0.02,),
+            Container(
+              width: MediaQuery.sizeOf(context).width * 0.25,
+              decoration: BoxDecoration(
+                color: Colors.blue.shade50,
+                //border: Border.all(color: Color(0xFF0F26A6), width:2),
+                borderRadius: BorderRadius.circular(16),
+              ),
+
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Image.asset("assets/images/cart-images/eco.png", width: 50,),
+                  SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
+                  Text("Eco Friendly", style: TextStyle(color: Color(0xFF0F26A6), fontSize: 10),),
+
+
+                ],
+              ),
             ),
 
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Image.asset("assets/images/cart-images/washing-machine.png", width: 60,),
-                SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
-                Text("Dry Cleaning", style: TextStyle(color: Color(0xFF0F26A6)),),
-
-
-              ],
-            ),
-          ),
-          SizedBox (width: MediaQuery.sizeOf(context).width * 0.05,),
-          Container(
-            width: MediaQuery.sizeOf(context).width * 0.36,
-            decoration: BoxDecoration(
-              color: Colors.blue.shade50,
-              //border: Border.all(color: Color(0xFF0F26A6), width:2),
-              borderRadius: BorderRadius.circular(16),
-            ),
-
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Image.asset("assets/images/cart-images/handwashing.png", width: 60,),
-                SizedBox(height: MediaQuery.sizeOf(context).height * 0.02,),
-                Text("Hand Washing", style: TextStyle(color: Color(0xFF0F26A6)),),
-
-
-              ],
-            ),
-          ),
-
-
-        ],
+          ],
+        ),
       ),
     );
   }
