+++ lib/features/cartc/widgets/fresh_start_full.dart
@@ -7,6 +7,33 @@ class FreshStartFull extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+          leading: Padding(
+            padding: const EdgeInsets.all(12.0),
+            child: GestureDetector(
+              onTap: () {
+                Navigator.pop(context);
+              },
+              child:
+              Container(
+                decoration: BoxDecoration(
+                  color: Theme.of(context).primaryColor,
+                  shape: BoxShape.circle,
+                ),
+                //padding: EdgeInsets.all(4),
+                child: Icon(Icons.arrow_back, color: Colors.white),
+              ),
+            ),
+          ),
+
+          title: Row(
+            children: [
+              //SizedBox(width: MediaQuery.of(context).size.width*0.05,),
+              Text("Popular Laundromat Nearby", style: (TextStyle(fontSize: 18)),),
+            ],
+          )
+
+      ),
       body: SafeArea(
         child: SingleChildScrollView(
           child: Padding(
@@ -14,16 +41,6 @@ class FreshStartFull extends StatelessWidget {
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Row(
-                  children: [
-                    GestureDetector(
-                      onTap: () {},
-                      child: Image.asset("assets/images/cart-images/circle-back.png", width: 30,),
-                    ),
-                    SizedBox(width: MediaQuery.of(context).size.width*0.05,),
-                    Text("Popular Laundromat Nearby", style: (TextStyle(fontSize: 18)),)
-                  ],
-                ),
                 SizedBox(height: MediaQuery.of(context).size.height*0.02,),
                 Image.asset("assets/images/cart-images/customer.png"),
                 SizedBox(height: MediaQuery.of(context).size.height*0.02,),
